+++ changelog.md
@@ -1,3 +1,7 @@
 ### Updated to MC 26.2
+* Added Sulfur Tiles (Stairs/Slab/Wall), Sulfur Pillar, Herringbone Sulfur Bricks & Sulfur Circular Paving
+* Added Cinnabar Tiles (Stairs/Slab/Wall), Cinnabar Pillar, Herringbone Cinnabar Bricks & Cinnabar Circular Paving
+* Added Polished Sulfur Pressure Plate & Button
+* Added Polished Cinnabar Pressure Plate & Button
 * Added/modified some tags for Sulfur Cube archetypes